204: Move PersistentHashMapBuilder tests
